requirements.txt
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.76
scikit-image==0.20.0
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.12.2
torch==2.0.1
torchvision==0.15.2
ftfy==6.1.1
regex==2023.6.3
tqdm==4.65.0
pillow==10.0.0
git+https://github.com/openai/CLIP.git
requests==2.31.0